+++ src/main/java/chess/controller/BoardController.java
@@ -1,15 +1,14 @@
-package chess.web;
+package chess.controller;
 
 import chess.service.GameService;
-import chess.web.dto.BoardDto;
-import chess.web.dto.CommendDto;
-import chess.web.dto.ResultDto;
+import chess.dto.BoardDto;
+import chess.dto.CommendDto;
+import chess.dto.ResultDto;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
