+++ src/main/java/chess/controller/HomeController.java
@@ -1,4 +1,4 @@
-package chess.web;
+package chess.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
