+++ src/main/java/chess/domain/Room.java
@@ -27,6 +27,12 @@ public class Room {
         }
     }
 
+    public void checkPassword(String password) {
+        if (!this.password.equals(password)) {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
+    }
+
     public String getName() {
         return name;
     }
