+++ src/main/java/chess/dto/CommendDto.java
@@ -1,4 +1,4 @@
-package chess.web.dto;
+package chess.dto;
 
 public class CommendDto {
