+++ src/main/java/chess/dto/RoomDto.java
@@ -1,4 +1,4 @@
-package chess.web.dto;
+package chess.dto;
 
 public class RoomDto {
